% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_sankey.R
\name{doi_sankey}
\alias{doi_sankey}
\title{<Add Title>}
\usage{
doi_sankey(values, edges, focus_node_id = NULL, width = NULL,
  height = NULL, size_max = 20, leaf_width = 10, leaf_height = 100)
}
\description{
<Add Description>
}

