% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_tree.R
\name{doi_tree}
\alias{doi_tree}
\title{<Add Title>}
\usage{
doi_tree(values, edges, focus_node_id = NULL, width = NULL, height = NULL,
  size_min = 0, size_max = 20, leaf_width = 10, leaf_height = 100)
}
\description{
<Add Description>
}

