% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{get_root}
\alias{get_root}
\title{Get root node}
\usage{
get_root(edges)
}
\arguments{
\item{el}{[character matrix] A matrix specifying the tree structure. The
first column are character names for the parents, the second are children.}
}
\value{
[character vector] A vector of nodes that have children but no
parents
}

