% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{merge_edge_values}
\alias{merge_edge_values}
\title{Merge edges and values into one data.table}
\usage{
merge_edge_values(edges, values)
}
\arguments{
\item{edges}{[matrix, data.frame, or data.table] Two columns of characters
giving parents and children for each edge in the tree.}

\item{values}{[data.frame or data.table] An object mapping unique units /
times / groups to their values. It must have a column "unit" giving the node
id, and optionally accepts columns "time" or "group" for timeboxes or DOI
sankeys.}
}
\value{
merged_values [data.table] A data.table mapping each edge to the
value for the associated child node.
}
\description{
Merge edges and values into one data.table
}

