% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{merge_timebox_display}
\alias{merge_timebox_display}
\title{Merge in default display for timebox trees / treeboxes}
\usage{
merge_timebox_display(opts)
}
\description{
Completes a partially filled list of display options.
}

