% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{merge_timebox_style}
\alias{merge_timebox_style}
\title{Merge in default style for timebox trees / treeboxes}
\usage{
merge_timebox_style(opts)
}
\description{
Completes a partially filled list of style options.
}

