% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timebox_tree.R
\name{timebox_tree}
\alias{timebox_tree}
\title{Create Timebox Trees}
\usage{
timebox_tree(values, edges, width = NULL, height = NULL,
  display_opts = list())
}
\description{
Timebox trees provide a way of linking time series and tree views, where
selections are made along the time series.
}
\examples{
data(sim_edges)
data(sim_data)
display_opts <- list("size_min" = 1)
timebox_tree(sim_data, sim_edges, display_opts = display_opts)
}

