% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{tip_descendants}
\alias{tip_descendants}
\title{Get all the descendants that are tips}
\usage{
tip_descendants(el, cur_node)
}
\arguments{
\item{el}{[character matrix] A matrix specifying the tree structure. The
first column are character names for the parents, the second are children.}

\item{cur_node}{[character] The name of the internal node to search for
descendants from. The name must appear in el.}
}
\value{
tips [character vector] A character vector of tip nodes that descend
from cur_node.
}
\description{
This is a helper for tree_sum, which gets the names of tip nodes
that descend from a specified internal node, given an edgelist specifying the
tree structure.
}
\examples{
library("phyloseq")
data(GlobalPatterns)
GP <- subset_taxa(GlobalPatterns, Phylum=="Chlamydiae")
el <- phy_tree(GP)$edge
tip_descendants(el, 34)
}

