% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{tree_fun_multi}
\alias{tree_fun_multi}
\title{Wrapper to compute tree aggregations over many samples}
\usage{
tree_fun_multi(el, values, tree_fun)
}
\arguments{
\item{el}{[character data.frame] The edgelist specifying the tree structure.
The first column are character names for the parents, the second are
children.}

\item{values}{[matrix] Tip values on which to aggregate. The column names
must be the same names as in el. Each row corresponds to a different sample.}

\item{tree_fun}{[function] The tree function to apply to each row of values.}
}
\value{
result [data.table] A data.table giving aggregated values for each of
sample, along with the row index of the sample on which that aggregation was
performed. The row index can be used for inputting sample / time / group
information after the fact.
}

