% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{tree_json}
\alias{tree_json}
\title{Generate a nested list representing a tree structure}
\usage{
tree_json(el, cur_node)
}
\arguments{
\item{el}{[data.frame] A data.frame with names "parent" and "child", giving
the edgelist for the tree structure to generate. Columns called "length" or
"depth" will also be accepted and stored in the resulting list.}

\item{cur_node}{The name of the root node in the edgelist, from which to
start the recursive descent.}
}
\value{
[list] A nested list whose structure reflects the edgelist in el. At
each level, "name" is a character string specifying the name of the root in
the subtree and children is a list with the same structure as the overall
tree. If "length" or "depth" are provided in "el", this information is stored
along with "name".
}
\description{
This representation is easily converted to the json required by
d3 by using toJSON in the jsonlite package.
}
\examples{
el <- data.frame(
  parent = c("1", "1", "2", "2", "3", "3"),
  child = c("2", "3", "4", "5", "6", "7"),
  stringsAsFactors = FALSE
)
tree_json(el, "1")
}

