% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{tree_sum}
\alias{tree_sum}
\title{Aggregate values in tips to internal nodes}
\usage{
tree_sum(el, values)
}
\arguments{
\item{el}{[character data.frame] The edgelist specifying the tree structure.
The first column are character names for the parents, the second are
children.}

\item{values}{[named vector] Tip values on which to aggregate. The names must
be the same names as in el.}
}
\description{
Given a value associated with each tip in a tree, this
calculates, for each internal node, the sum across all tips that descend from
it.
}
\examples{
el <- data.frame(
  parent = c("1", "1","2", "2", "2"),
  child = c("2", "7", "3", "4", "5"),
  stringsAsFactors = FALSE
)
counts  <- c(
  "7" = 10,
  "3" = 2,
  "4" = 5,
  "5" = 1
)
tree_sum(el, counts)
}

